# Conflicts:
#	hatch/src/project.rs
#	hatch/src/task.rs

+++ hatch/src/assets/tests/generator.rs
@@ -8,38 +8,53 @@ use std::path::PathBuf;
 fn generate_one_without_directories() {
   let test_asset = ProjectAsset::new(PathBuf::from("./"), String::from("test.test"), String::from("test"));
 
-  generate_one(&test_asset);
+  if let Err(e) = generate_one(&test_asset) {
+    panic!(e);
+  }
 
   match fs::File::open("./test.test") {
     Ok(mut file) => {
       let mut contents = String::new();
-      file.read_to_string(&mut contents);
+      let result = file.read_to_string(&mut contents);
+      fs::remove_file("./test.test");
+
+      if let Err(e) = result {
+        panic!(e)
+      }
     
       assert_eq!(contents, "test");
-      
-      fs::remove_file("./test.test");
     },
-    Err(e) => panic!(e)
+    Err(e) => {
+      fs::remove_file("./test.test");
+      panic!(e)
+    }
   }
 }
 
 #[test]
 fn generate_one_with_directories() {
-  let test_asset = ProjectAsset::new(PathBuf::from("./test/"), String::from("test.test"), String::from("test"));
+  let test_asset = ProjectAsset::new(PathBuf::from("./foo/"), String::from("test.test"), String::from("test"));
 
-  generate_one(&test_asset);
+  if let Err(e) = generate_one(&test_asset) {
+    panic!(e);
+  }
 
-  match fs::File::open("./test/test.test") {
+  match fs::File::open("./foo/test.test") {
     Ok(mut file) => {
       let mut contents = String::new();
-      file.read_to_string(&mut contents);
-    
-      assert_eq!(contents, "test");
-      
-      fs::remove_file("./test/test.test");
-      fs::remove_dir("./test/");
+      let result = file.read_to_string(&mut contents);
+      fs::remove_file("./foo/test.test");
+      fs::remove_dir("./foo/");
+
+      if let Err(e) = result {
+        panic!(e)
+      }
     },
-    Err(e) => panic!(e)
+    Err(e) => {
+      fs::remove_file("./foo/test.test");
+      fs::remove_dir("./foo/");
+      panic!(e)
+    }
   }
 }
 
@@ -47,34 +62,53 @@ fn generate_one_with_directories() {
 fn generate_one_overwrites_file() {
   let test_asset = ProjectAsset::new(PathBuf::from("./"), String::from("test2.test"), String::from("old"));
 
-  generate_one(&test_asset);
+  if let Err(e) = generate_one(&test_asset) {
+    panic!(e);
+  }
 
   // verify that the old content is there
   match fs::File::open("./test2.test") {
     Ok(mut file) => {
       let mut contents = String::new();
-      file.read_to_string(&mut contents);
-    
+      let result = file.read_to_string(&mut contents);
+      fs::remove_file("./test2.test");
+
+      if let Err(e) = result {
+        panic!(e)
+      }
+
       assert_eq!(contents, "old");
     },
-    Err(e) => panic!(e)
+    Err(e) => {
+      fs::remove_file("./test2.test");
+      panic!(e)
+    }
   }
 
   let test_asset = ProjectAsset::new(PathBuf::from("./"), String::from("test2.test"), String::from("new"));
 
-  generate_one(&test_asset);
+  if let Err(e) = generate_one(&test_asset) {
+    fs::remove_file("./test2.test");
+    panic!(e);
+  }
 
   // verify that the new content overwrites the old content
   match fs::File::open("./test2.test") {
     Ok(mut file) => {
       let mut contents = String::new();
-      file.read_to_string(&mut contents);
+      let result = file.read_to_string(&mut contents);
+      fs::remove_file("./test2.test");
+
+      if let Err(e) = result {
+        panic!(e)
+      }
     
       assert_eq!(contents, "new");
-      
-      fs::remove_file("./test2.test");
     },
-    Err(e) => panic!(e)
+    Err(e) => {
+      fs::remove_file("./test2.test");
+      panic!(e)
+    }
   }
 }
 
@@ -85,29 +119,47 @@ fn generate_all_assets() {
 
   let test_assets = vec![test_asset_one, test_asset_two];
 
-  generate_all(&test_assets);
+  if let Err(e) = generate_all(&test_assets) {
+    fs::remove_file("./one.test");
+    fs::remove_file("./two.test");
+    panic!(e);
+  }
 
   match fs::File::open("./one.test") {
     Ok(mut file) => {
       let mut contents = String::new();
-      file.read_to_string(&mut contents);
+      let result = file.read_to_string(&mut contents);
+      fs::remove_file("./one.test");
+
+      if let Err(e) = result {
+        fs::remove_file("./two.test");
+        panic!(e)
+      }
     
       assert_eq!(contents, "one");
-      
-      fs::remove_file("./one.test");
     },
-    Err(e) => panic!(e)
+    Err(e) => {
+      fs::remove_file("./one.test");
+      fs::remove_file("./two.test");
+      panic!(e)
+    }
   }
 
     match fs::File::open("./two.test") {
     Ok(mut file) => {
       let mut contents = String::new();
-      file.read_to_string(&mut contents);
+      let result = file.read_to_string(&mut contents);
+      fs::remove_file("./two.test");
+
+      if let Err(e) = result {
+        panic!(e)
+      }
     
       assert_eq!(contents, "two");
-      
-      fs::remove_file("./two.test");
     },
-    Err(e) => panic!(e)
+    Err(e) => {
+      fs::remove_file("./two.test");
+      panic!(e)
+    }
   }
 }
